requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
mysqlclient